requirements.txt
{\rtf1\ansi\ansicpg1251\cocoartf2822
\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fnil\fcharset0 .AppleSystemUIFontMonospaced-Regular;}
{\colortbl;\red255\green255\blue255;\red0\green0\blue0;\red20\green0\blue196;}
{\*\expandedcolortbl;;\csgray\c0;\cssrgb\c10980\c0\c81176;}
\paperw11900\paperh16840\margl1440\margr1440\vieww30040\viewh17760\viewkind0
\pard\tx560\tx1120\tx1680\tx2240\tx2800\tx3360\tx3920\tx4480\tx5040\tx5600\tx6160\tx6720\pardirnatural\partightenfactor0

\f0\fs26 \cf2 python-telegram-bot==\cf3 20.7\cf2 \
gspread\
oauth2client\
python-dotenv\
nest_asyncio}